requirements.txt
asgiref==3.3.1
Django==3.1.5
pytz==2020.5
sqlparse==0.4.1
djangorestframework
psycopg2-binary>=2.8
requests
djangorestframework-jwt
django-q